[QOL-9055] break circular dependency between 'tasks' and 'lib'
- move enqueueing functions to 'tasks'
